Every dao id is now integer & plase use the Integer wrapper

+++ src/main/webapp/admin/controller/trainee/DeactivateTrainee.java
@@ -21,7 +21,7 @@ public class DeactivateTrainee extends HttpServlet {
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        long id = Long.parseLong(request.getParameter("id"));
+        int id = Integer.parseInt(request.getParameter("id"));
 
         TraineeDao traineeDao = DAOFactory.instance(DAOFactory.HIBERNATE).getTraineeDAO();
         Trainee trainee = traineeDao.findById(id, false);
